+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Handled duplicate lesson uploads by returning a `409` Conflict
 - Added `deleted_lesson.rb` model and ability to hard delete lessons
 - Cleanup and improved performance when viewing lessons